+++ src/lib.rs
@@ -89,6 +89,7 @@ pub mod json;
 pub mod rust;
 #[doc(hidden)]
 pub mod with_prefix;
+mod duplicate_key_impls;
 
 /// Separator for string-based collection de/serialization
 pub trait Separator {